requirements.txt
streamlit
paho-mqtt
pandas
scikit-learn
joblib
numpy